security-cam/requirements.txt
python-socketio==5.0.4
aiohttp
async-timeout
asyncio
scikit-image